02-notifier/requirements.txt
aiogram==3.7.0
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
